Changed sample count value to unsinged 32 bit so we can average more samples. Calibrated the prototype board a little.

+++ hardware/firmware/remoteMonitorFirmware/protocol.h
@@ -44,7 +44,7 @@ typedef struct __attribute__((packed)) {
 	packetType_t type;
 	uint8_t requestSequence;
 	uint8_t dataSequence;
-	uint16_t sampleCount;
+	uint32_t sampleCount;
 	int64_t powerData;
 	#ifdef EXTENDED_DATA_PACKET
 	int16_t linePeriod;
